🔨 Forge: Implement Token Burn Rate Forecasting and SQLite Throttling
This PR implements the requested Token Burn Rate Forecasting Engine and Standalone Concurrency Throttling.

Key changes:
- Updated `Forecaster` to use Exponentially Weighted Moving Average (EWMA) for accurate token burn rate predictions per tenant in `srcs/server/telemetry/forecaster.go`.
- Implemented comprehensive SQLite concurrency throttling in `SIPDB` (`srcs/server/orchestration/sip.go`) using a semaphore to prevent `database is locked` errors in Standalone mode.
- Added a new Grafana dashboard panel configuration for 24h token burn rate prediction in `srcs/server/monitoring/dashboards/token_forecast.json`.
- Fixed type mismatch issues in `srcs/server/agents/builtin_rust_integration_test/integration_test.go` discovered during CI.
- Restored integrity of Bazel BUILD files in `srcs/proto/` and `srcs/server/agents/builtin_rust/`.
- Added unit tests for the EWMA forecasting logic in `srcs/server/telemetry/forecaster_test.go`.

Verified with `bazelisk test` for orchestration and telemetry packages.

issue_id: 5654

+++ srcs/server/agents/builtin_rust_integration_test/integration_test.go
@@ -213,12 +213,12 @@ func TestRustAgent_RunTask_NoLLM(t *testing.T) {
 			break
 		}
 		t.Logf("event: type=%v", evt.Type)
-		if evt.Type == int32(agentservicepb.EventType_RUN_STARTED) {
+		if evt.Type == agentservicepb.EventType_RUN_STARTED {
 			sawRunStarted = true
 		}
 		// TASK_ERROR or TASK_COMPLETE terminate the stream.
-		if evt.Type == int32(agentservicepb.EventType_TASK_ERROR) ||
-			evt.Type == int32(agentservicepb.EventType_TASK_COMPLETE) {
+		if evt.Type == agentservicepb.EventType_TASK_ERROR ||
+			evt.Type == agentservicepb.EventType_TASK_COMPLETE {
 			break
 		}
 	}
